Installation directory (this will create a google-cloud-sdk subdirectory) /home/codespace install_dir/google-cloud-sdk/platform/gsutil/gslib/vendored/boto/tests/unit/cloudfront/test_distribution.py
import unittest

from boto.cloudfront.distribution import DistributionConfig
from boto.cloudfront.logging import LoggingInfo


class CloudfrontDistributionTest(unittest.TestCase):
    cloudfront = True

    def setUp(self):
        self.dist = DistributionConfig()

    def test_logging(self):
        # Default.
        self.assertEqual(self.dist.logging, None)

        # Override.
        lo = LoggingInfo(bucket='whatever', prefix='override_')
        dist = DistributionConfig(logging=lo)
        self.assertEqual(dist.logging.bucket, 'whatever')
        self.assertEqual(dist.logging.prefix, 'override_')
